First version with all buttons wired. This is for DFPlayer Pro but none of the audio functions are implemented.
This also fully deals with conversion between Teensy C code and Wiring (Arduino) code.
There is some strange bug in the libraries: hopefully this is now resolved.

You need to pick the kind of USB usage in the Arduino menu (e.g. Keyboard). Very little
code would need to be added to make this a USB keyboard.